Patch Schulze STV memory leaks, preliminary QPQ port, fix int/size_t stuff.
The QPQ port imports a bunch of functions I shouldn't need for tie
breaking; I'll fix that later.

+++ src/multiwinner/compat_qbuck.cc
@@ -260,7 +260,7 @@ std::vector<int> old_qltd_pr::QuotaBucklin(
 			ordering social_order = borda(PT_WHOLE).elect(ballots,
 					hopefuls, num_candidates);
 
-			std::vector<int> winners = ordering_tools::get_winners(
+			std::vector<size_t> winners = ordering_tools::get_winners(
 					social_order);
 
 			// Pick a random winner.
